restore sanity to robot.py
